项目成本核算编辑

+++ internal/app/model/impl/gorm/internal/entity/e_tax_calculation.go
@@ -32,13 +32,13 @@ func (a SchemaTaxCalculation) ToTaxCalculation() *TaxCalculation {
 // TaxCalculation 税目计算表实体
 type TaxCalculation struct {
 	CostModel
-	RecordID           *string `gorm:"column:record_id;size:36;index;"`      // 记录ID
-	Category           *string `gorm:"column:category;size:50;index;"`       // 类别
-	Name               *string `gorm:"column:name;size:200;index;"`          // 税目名称
-	CalculationFormula *string `gorm:"column:calculation_formula;size:200;"` // 计算公式
-	Type               *int    `gorm:"column:type;index;"`                   // 税率类型(1:含税计算 2:不含税计算)
-	TaxRate            *int    `gorm:"column:tax_rate;"`                     // 税率
-	Memo               *string `gorm:"column:memo;size:1024;"`               // 备注
+	RecordID           *string  `gorm:"column:record_id;size:36;index;"`      // 记录ID
+	Category           *string  `gorm:"column:category;size:50;index;"`       // 类别
+	Name               *string  `gorm:"column:name;size:200;index;"`          // 税目名称
+	CalculationFormula *string  `gorm:"column:calculation_formula;size:200;"` // 计算公式
+	Type               *int     `gorm:"column:type;index;"`                   // 税率类型(1:含税计算 2:不含税计算)
+	TaxRate            *float64 `gorm:"column:tax_rate;"`                     // 税率
+	Memo               *string  `gorm:"column:memo;size:1024;"`               // 备注
 }
 
 func (a TaxCalculation) String() string {
